Add links in README to the various products
It'd be useful in the blurb about what APIs the project contains
to link to where the source for these features live in the project.

I'm not too worried about link rot, as while the APIs in the products
may change over time, the product names I'd imagine will stay static.