LinuxContainer: Ensure we close guest agent connections
Use scoped access to the agent connection to ensure the vsock fds
are close(2)'d. In a follow up I'd like to change LinuxProcess to
dialing the agent connection inside of `.start()`. I don't like
needing to pass it an agent connection. It has a reference to a vm
so it can easily dial itself if it needs to.