LinuxContainer: Add run method
To cut down on some of the boilerplate for running
containers we can have a run method which is common
on a lot of process wrappers. This method does
create + start + wait for you, and optionally cleans
up (stops) the container as well if you provide true
for the only argument named `stop`.