ContainerizationOCI: Cut down on allocations
We were doing intermediate step copies to Data objects
both for push and pull. We only need the data for the lifetime
of the singular writes to update the state of the checksums and
to write to disk somewhere, so we can use a view into the buffer
from the http client to satisfy this.