Vminitd: Rework ManagedContainer
This gets rid of some design decisions that I was not fond of.
This changes the process start for both execs and the init process
to happen internally in the object, as well as makes it so that all
exec functionalities can happen on the container object itself. This
allows us to get rid of the `if request.id == containerID` branches
in the rpcs, and handle this entirely in the object itself using its
own internal state.