VZVirtualMachineInstance: Rework installRosetta flow
Because we were installing rosetta directly in the constructor for
VZVirtualMachineInstance, and we'd prefer to not have the constructor
async as it pollutes so much more, we had devised this gnarly callback
approach for the install that is a bit of an eyesore. This changes focus
is on moving the install flow to a method that is already async so we
can piggyback off of it, and removing the install logic from the
config -> VZConfig conversion. Now the install will occur during start()
if rosetta is not installed.