Vminitd: Rework ManagedProcess.IO protocols
I think the way these types were structured made
reasoning about them a bit difficult. I don't think
this fully solves the problem, but this change aims to
make things a bit simpler by hoisting a lot of the logic
for the IO relays to a new IOPair type thats goal is to
simply take in a reader and writer and handle their resource
cleanup after a relay finishes. Bundled in with this is also
the buffer we'll use to copy between them.

This change:

- Alters ManagedProcess.IO `start` to take in the process to alter
instead of this just being a side effect of the constructors.
- Gets rid of `close()` in favor of `CloseStdin`. The IO will get closed
when the relays finish, which will naturally happen if the process exits
or just closes its side of the pipes/pty. This makes it so that the one
special case (a client wants to signal no more input is coming) is still
sane.
- Move all relay logic and resource cleanup to a new IOPair type that takes
in protocols that are easily conformable by all of our various io types (
Socket, Terminal, FileHandle).