vminitd: Proper pty setup/mount /dev/console
We need to allocate the pty in the containers ns
otherwise you run into funny outcomes later. This
change additionally mounts /dev/console by binding
the pty opened in the containers namespace onto it.
Today /dev/console is solely from devtmpfs which
points at /dev/hvc0, or the virtio serial console.
This causes systemds output to go to serial/the bootlog
instead of what folks would expect. This should make
systemd like us a lot more than it does today.

Another large change that was somewhat required here is
the changes in IOPair. EPOLLHUP on a pty is deceptively
weird to handle as it could be reopened, so we can't do
what we were doing today and just close all the fds.
To remedy this I skip exiting on HUP for pty's and
instead just cleanup at process exit time, but ensure
to drain the pty beforehand.

Co-authored-by: Elijah Wright <git@elijahs.space>