ContainerizationOCI: Add OCI prefix to types
A large majority of the types in this package were
extremely generically named making it common to have to
reference the types as ContainerizationOCI.TheType.
This change prefixes all of the core runtime spec, and
image spec types with OCI, and then adjusts our code to
reference these new names so we still build :)

This is a breaking change, but I'd rather pull the bandaid
now.