LinuxContainer/LinuxProcess: Rework supplying configuration
This is a fairly large reworking, but it gets rid of something
that has plagued this since release which is the properties
needing to be locked to be Sendable compliant. This was somewhat
of a copout because we mostly know there's not a great deal of
ways to have misused the setup today, but alas we'd need to either
mark the type as @unchecked or just find a different route for
setting the configuration. This change:

Exposes the underlying Configuration type that today only housed
things that aren't on the OCI spec. I'd love to just expose the OCI
spec, but we don't (and possibly never will) support everything on the
spec, so exposing it to be freely modified would be a bit odd. Now
everything related to the container is configured on this type,
and the same goes for execs.