LinuxContainer: Allow reuse after being stopped
After the container is stopped it's a bit odd how we don't allow
the object to be restarted. If someone wanted to continuously
rerun the same container they'd need to make a new object every time
even though all guest state is blown away on stop() so it's a clean
slate.

This change makes it so that you can create+start again after stop
(although create stands out like a sore thumb now as stop -> create
is just strange).