Allow reusing virtiofs mounts
Fwiw, this is a bit odd, but it's also very simple
to support and feels odd that this was failing today.
This also adds in a tiny bit of logic in hashMountSource
to resolve the path of symlinks, so now even symlinks
to the same dir can be supplied N times without a new
virtiofs device.