Agent: Add and implement writeFile
This change adds in a new rpc that we probably should swap to
for the underlying implementation of a couple rpcs we currently have.
This new rpc simply writes to an existing or new file some data.
This is useful for guest (and container) setup for things like /etc/hosts,
/etc/resolv.conf, writing to one-off cgroup files and so on.

In this change currently I've just implemented subtree_control toggling
for the root cg in standardSetup with this new change, but in the future
I'd like to swap the implementations of our /etc/hosts and /etc/resolv.conf
logic with this.