vminitd: Add init and execs to cgroup
We currently weren't doing any cgroup setup whatsoever.
For the most part this doesn't matter too much, however
certain images that may fool around with cgroups don't
like this :). Lets do the bare minimum these expect which
is to at least have the process running in a nested cg
and not apart of the root cg.