vminitd: Wait for execvpe to return to continue
Today, because we don't wait for execvpe to finish to continue onwards,
it's possible that if you did an exec quick enough after starting an init
process for a container, that you could join the init processes namespaces
before pivot_root has taken place which is quite fun. Let's wait for exec to
finish (or an error to occur) to prevent this.