ContainerizationOS: Rework User type
There's a couple things we can add to the User types to make them more
friendly.

1. We should have errors specific to this type so someone can catch them and
take action (e.g. the user didn't exist in the file, or the file itself doesn't
exist).
2. ExecUser should expose the shell if there is one.
3. getExecUser (rename of parsePasswd) should have smarter fallbacks in some spots.

The largest change here is I've essentially reworked getExecUser to be much more
close to runc's logic, as that's what most folks are expecting.