LinuxContainer/LinuxProcess: Have wait return ExitStatus
Have wait return a struct containing the exit code and the time
at which the process actually exited. This is useful so higher level
tools (like container) don't need to craft a date themselves, and it
will be more accurate as it's taken in the guest directly after the
process exits.