Kernel: Fix `make`
Today the kernel subdirectory is somewhat useless. The dockerfile
we had tried to use an image that isn't actually on ghcr, so
everything would kind of just crap out. This change swaps things
so that we build a local image that has all the bits needed to
compile the kernel, and then uses that image to build the resulting
kernel and plop it in the current working directory. This change
does add a reliance on `container` being installed however, but
for folks that need to build a custom kernel this doesn't seem
like too much of an issue to me.