Vminitd: Write cgroup limits
Today we don't actually setup any cgroup limits, as because there's
a 1-1 mapping from container<->vm we can just use the VMs resources
as the limit (can't use more than 1GB if that's all the guest sees :) ).
However, if we ever supported > 1 container in the guest it'd be necessary
to actually setup the cg limits. This change just sets a memory limit and cpu
toggles to match whatever was specific for the container.