Introduce new experimental LinuxPod type
Introduces a new type capable of running > 1 container in the guest.
The API mostly follows LinuxContainer, and each individual container
can be addressed via any of the methods that require you to pass
the containerID as the first param. Today there's no support for
namespace sharing, but that shouldn't be terrible to support.