LinuxContainer: Add bootlog as configuration field
Previously, the bootlog was supplied once in the constructor to
VZVirtualMachineManager which meant that if you used this same
manager for multiple ctrs that all logs would end up going to
the same file, which becomes quite cumbersome to follow..

This change moves bootlog to be a container configuration param
and also moves it to be a VMConfiguration param, so it can be
threaded through from LinuxContainer -> vmm.create() and be truly
container unique now. The largest driver for this was the integration
tests which today every single test spits out logs to a singular file,
making guest investigations tricky to actually look into.