Upgrade to GRPC V2
No real notable API changes. The largest change code wise is in Vminitd.
The connection loop is setup immediately in the new API by calling
runConnections() on the client, and this blocks until either the task
it's running on is cancelled or beginGracefulShutdown is called. Because
of this, either we'd have to modify our API to have VirtualMachineAgent
implementations call some run() method, or we just put runConnections in
an internal task on Vminitd. I went for the latter approach as it's simpler.