Chore: Use guestVsockPorts
This was a funny oversight. We have two port spaces, one for listening
sockets on the host, and one for vsock proxies in the guest, but only
the host "allocator" was being used. This didn't really matter as
the ports in the guest would still be unique, but would still be good
to fix.