EXT4: Optimize unpack
Optimize unpack a little by trying to reduce allocations in the hot
path. Today for every file we read the entire file into memory and
then pass the data blob to the ext4 writer to eventually be written
to the sparse file. Before being written to the sparse file the data
is copied *again* to a temp buffer before finally hitting write(2)
in FileHandle.

This change moves things around such that we can pass an optional
buffer to the ext4 create() (so we can reuse a buffer for file writes),
as well as stops reading entire files into memory by passing the archive
entry itself (wrapped in a ReaderStream object albeit) down to the
writer.