Don't ignore possible errors from dup(2)
Everytime we grab a vsock connection we dup the conn and close
the original, otherwise we'd need to carry around the vsock connection
type everywhere as it closes the fd in its destructor. We weren't checking
the return value of dup however, so if it did fail we'd have a useless
filehandle with an fd of -1.