Integration: Add stdio ingest test
Add a test to ingest 500MiB of data across stdout/stderr. Useful to test
how fast we can ingest stdio. The timing will always be a tad off as it
times between start and wait returning, and there's quite a lot in the way..
but it's a good enough metric.