LinuxContainer: Keep reference to vended execs
This change is aimed at making forgetting to call .delete()
on a LinuxProcess less destructive than it can be. Because
Virt.framework invalidates any vsock fds it vended if the vm
is stopped, trying to perform some operations on the grpc client
through any of the process methods could trigger an ebadf, which NIO
asserts on. This keeps a reference to the execs and deletes all of them
for you once the container dies. I still think leaving .delete a public
method is useful as otherwise the stdio fds are left open, but cleanup
should occur all in one place now if you don't care about this.

This additionally:

1. Fixes two of our tests that forgot to delete() an exec.
2. Adds two new tests to verify that process.delete() is now idempotent,
and we don't need to call delete().