Add new FileHandle option for serial console output
Taking in a filehandle gives the user quite a bit more freedom on how to handle boot log output. They can set up a
kqueue watch on it and redirect output somewhere else etc etc. The implementation for this has us take in a new BootLog
type that has two options:

1. .file, which is analogous to what we had prior. Just provide a URL and a true by default append
field.
2. .fileHandle which is the new addition. Can pass any fd that is writable, and the VMM should write serial
console output to it.