EXT4: Remove advertising inline data
We've seen issues arising from us advertising inlinedata on
our inodes. It doesn't seem like we actually do support this
at all, I think it was a mistake in mixing up inline xattrs and
data. So, stop advertising this feature on the superblock, and
stop setting the flag on inodes.