Cgroup2Manager: Fix cgroup deletions
If there's any nested cgroups in the one we made for the container
(commonly seen for systemd images) removeItem didn't seem to be
having a grand time, even though it states it should do recursive
removals. Lets roll our own, and have a small EBUSY/EAGAIN retry loop
as well. This fixes LinuxContainer.stop() for any containers with nested
cgs.