Support read only rootfs
We sort of do today, unless you ask for dns/host settings. The problem
is we can't write the files if the rootfs is ro. What other runtimes do
is to delay the ro application, bind mount in some host/resolv.conf files
they made, and then the OCI runtime eventually remounts the rootfs ro. We
can do the same here.